python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "entrance"
version = "1.2.0"
author = "Ensoft Ltd"
description = "Server framerwork for web apps"
readme = "README.md"
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml",
    "uvicorn",
    "fastapi",
    "websockets",
    "ujson",
]

[project.extra-dependencies]
with-router-features = [
    "janus",
    "ncclient",
    "paramiko",
]

[project.urls]
Repository = "https://github.com/ensoft/entrance"
Issues = "https://github.com/ensoft/entrance/issues"

[tool.hatch.build]
include = ["entrance"]

[tool.isort]
profile = "black"
line_length = 79

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".pyenv",
    ".ruff_cache",
    ".vscode",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 79
indent-width = 4

# Assume Python 3.9 (minimum supported version)
target-version = "py39"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Disable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
docstring-code-format = false
